+++ mayan/apps/autoadmin/literals.py
@@ -1,0 +1,5 @@
+from __future__ import unicode_literals
+
+DEFAULT_EMAIL = 'autoadmin@example.com'
+DEFAULT_PASSWORD = None
+DEFAULT_USERNAME = 'admin'