+++ mayan/apps/autoadmin/tests/literals.py
@@ -1,0 +1,7 @@
+from __future__ import unicode_literals
+
+TEST_ADMIN_USER_EMAIL = 'testemail@example.com'
+TEST_ADMIN_USER_PASSWORD = 'test admin user password'
+TEST_ADMIN_USER_USERNAME = 'test_admin_user_username'
+TEST_FIRST_TIME_LOGIN_TEXT = 'First time login'
+TEST_MOCK_VIEW_TEXT = 'mock view text'